gbm: Add map/unmap functions
This adds map and unmap functions to GBM utilizing the DRIimage extension
mapImage/unmapImage functions or existing internal mapping for dumb
buffers. Unlike prior attempts, this version provides a region to map and
usage flags for the mapping. The operation follows the same semantics as
the gallium transfer_map() function.

This was tested with GBM based gralloc on Android.

This still creates a context, but I've moved it into gbm_create_device
rather than in the map function. This should remove any need for reference
counting and problems with memory leaks.

Signed-off-by: Rob Herring <robh@kernel.org>